room_booksコントローラーを作成し、bookコントローラーからroom内のbook操作に関する記述を切り出し
