+++ app/helpers/room_books_helper.rb
@@ -1,0 +1,2 @@
+module RoomBooksHelper
+end
