+++ app/views/shared/_rooms_sidebar.html.erb
@@ -35,7 +35,7 @@
     <div class=" d-flex fw-bold pt-3 mb-3 partition-border"><span class="text-danger">N</span>ew Book</div>
     <div class="room-new-book">このRoomで新しいBookを作る</div>
     <div class="room-new-book-btn">
-      <%= link_to "作成", new_room_book_path(@room.id), class:"room-new-book-link" %>
+      <%= link_to "作成", new_room_room_book_path(@room.id), class:"room-new-book-link" %>
     </div>
     <%# メンバー管理 %>
     <div class="fw-bold pt-3 mb-3 partition-border"><span class="text-danger">M</span>embers</div>
