+++ config/routes.rb
@@ -4,7 +4,7 @@ Rails.application.routes.draw do
   root to: "users#index"
   resources :users, only: [:update]
   resources :rooms, only: [:new, :create, :show] do
-    resources :books, only: [:new, :create, :update]
+    resources :room_books, only: [:new, :create, :update]
     resources :chat_messages, only: [:index, :create]
     collection do
       post "add_member_to_lists"
