+++ spec/requests/room_books_request_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe "RoomBooks", type: :request do
+
+end
